consume delete endpoints

+++ UI/delete.js
@@ -1,13 +1,18 @@
-// edit record fetch method
-let deleteRecords = document.getElementsByClassName('delete_record_btn');
-let ids = document.getElementsByClassName('record_id');
+// delete record fetch method
+let records = document.getElementById('past_records');
 
-for(let i = 0; i < deleteRecords.length; i++) {
-    deleteRecords[i].addEventListener("click", function deleteRecord(event){
-    event.preventDefault();
-    record_id = ids[i].innerText;
+records.addEventListener("click", function(event) {
+	// event.target is the clicked element!
+	// If it was a button
+	if (event.target.className != 'delete_record_btn') return;
+
+    let button = event.target.closest('.individual_record');
+   
     
-    fetch(`https://ireporter-valerie.herokuapp.com/api/v2/incidents/${record_id}`, {
+    // button found!  retrieve the ID and use it
+    event.preventDefault();
+
+    fetch(`https://ireporter-valerie.herokuapp.com/api/v2/incidents/${event.target.id}`, {
                 method: 'DELETE',
                 headers: {
                     'Authorization': bearer,
@@ -18,14 +23,15 @@ for(let i = 0; i < deleteRecords.length; i++) {
             })
             .then((response) => response.json())
             .then((data) => {
-                if(data.status === 201){
-                    console.log(data)
-                }else{
-                    console.log(data)
-                    window.alert(JSON.stringify(data['message']))
-                }
+                console.log(data)
+                // if(data.status === 201){
+                //     console.log(data)
+                // }else{
+                //     console.log(data)
+                //     window.alert(JSON.stringify(data['message']))
+                // }
             })
-    
-})
-}
 
+    // button.remove();
+		
+})
